+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- [browser] feat: Add build bundle including @sentry/apm
+- [apm] feat: Add build bundle including @sentry/browser + @sentry/apm
 - [utils] ref: Extract adding source context incl. tests
 
 ## 5.11.0